feat: picture component moved to hoppscotch-ui (#3032)
